+++ cpp/src/types.cpp
@@ -173,10 +173,10 @@ Input::Input(std::vector<int64_t> shape, TensorFormat format) {
   this->dtype = DataType::kUnknown;
   this->format = format;
   this->input_is_dynamic = false;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
-Input::Input(std::vector<int64_t> shape, std::vector<int64_t> tensor_domain, TensorFormat format) {
+Input::Input(std::vector<int64_t> shape, std::vector<double> tensor_domain, TensorFormat format) {
   this->opt_shape = shape;
   this->min_shape = shape;
   this->max_shape = shape;
@@ -195,10 +195,10 @@ Input::Input(std::vector<int64_t> shape, DataType dtype, TensorFormat format) {
   this->dtype = dtype;
   this->format = format;
   this->input_is_dynamic = false;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
-Input::Input(std::vector<int64_t> shape, DataType dtype, std::vector<int64_t> tensor_domain, TensorFormat format) {
+Input::Input(std::vector<int64_t> shape, DataType dtype, std::vector<double> tensor_domain, TensorFormat format) {
   this->opt_shape = shape;
   this->min_shape = shape;
   this->max_shape = shape;
@@ -217,10 +217,10 @@ Input::Input(c10::IntArrayRef shape, TensorFormat format) {
   this->dtype = DataType::kUnknown;
   this->format = format;
   this->input_is_dynamic = false;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
-Input::Input(c10::IntArrayRef shape, std::vector<int64_t> tensor_domain, TensorFormat format) {
+Input::Input(c10::IntArrayRef shape, std::vector<double> tensor_domain, TensorFormat format) {
   this->opt_shape = torch_tensorrt::core::util::toVec(shape);
   this->min_shape = torch_tensorrt::core::util::toVec(shape);
   this->max_shape = torch_tensorrt::core::util::toVec(shape);
@@ -239,10 +239,10 @@ Input::Input(c10::IntArrayRef shape, DataType dtype, TensorFormat format) {
   this->dtype = dtype;
   this->format = format;
   this->input_is_dynamic = false;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
-Input::Input(c10::IntArrayRef shape, DataType dtype, std::vector<int64_t> tensor_domain, TensorFormat format) {
+Input::Input(c10::IntArrayRef shape, DataType dtype, std::vector<double> tensor_domain, TensorFormat format) {
   this->opt_shape = torch_tensorrt::core::util::toVec(shape);
   this->min_shape = torch_tensorrt::core::util::toVec(shape);
   this->max_shape = torch_tensorrt::core::util::toVec(shape);
@@ -266,14 +266,14 @@ Input::Input(
   this->dtype = DataType::kUnknown;
   this->format = format;
   this->input_is_dynamic = true;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
 Input::Input(
     std::vector<int64_t> min_shape,
     std::vector<int64_t> opt_shape,
     std::vector<int64_t> max_shape,
-    std::vector<int64_t> tensor_domain,
+    std::vector<double> tensor_domain,
     TensorFormat format) {
   this->opt_shape = opt_shape;
   this->min_shape = min_shape;
@@ -300,7 +300,7 @@ Input::Input(
   this->dtype = dtype;
   this->format = format;
   this->input_is_dynamic = true;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
 Input::Input(
@@ -308,7 +308,7 @@ Input::Input(
     std::vector<int64_t> opt_shape,
     std::vector<int64_t> max_shape,
     DataType dtype,
-    std::vector<int64_t> tensor_domain,
+    std::vector<double> tensor_domain,
     TensorFormat format) {
   this->opt_shape = opt_shape;
   this->min_shape = min_shape;
@@ -330,14 +330,14 @@ Input::Input(c10::IntArrayRef min_shape, c10::IntArrayRef opt_shape, c10::IntArr
   this->dtype = DataType::kUnknown;
   this->format = format;
   this->input_is_dynamic = true;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
 Input::Input(
     c10::IntArrayRef min_shape,
     c10::IntArrayRef opt_shape,
     c10::IntArrayRef max_shape,
-    std::vector<int64_t> tensor_domain,
+    std::vector<double> tensor_domain,
     TensorFormat format) {
   this->opt_shape = torch_tensorrt::core::util::toVec(opt_shape);
   this->min_shape = torch_tensorrt::core::util::toVec(min_shape);
@@ -364,7 +364,7 @@ Input::Input(
   this->dtype = dtype;
   this->format = format;
   this->input_is_dynamic = true;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
 Input::Input(
@@ -372,7 +372,7 @@ Input::Input(
     c10::IntArrayRef opt_shape,
     c10::IntArrayRef max_shape,
     DataType dtype,
-    std::vector<int64_t> tensor_domain,
+    std::vector<double> tensor_domain,
     TensorFormat format) {
   this->opt_shape = torch_tensorrt::core::util::toVec(opt_shape);
   this->min_shape = torch_tensorrt::core::util::toVec(min_shape);
@@ -402,7 +402,7 @@ Input::Input(at::Tensor tensor) {
   }
   this->format = frmt;
   this->input_is_dynamic = false;
-  this->tensor_domain = std::vector<int64_t>{0, 2};
+  this->tensor_domain = std::vector<double>{0, 2};
 }
 
 /* ==========================================*/
